glib: Move ControlFlow to it own file
As it will be shared between sources and signals
Continue & Inhibit previously
